refactor: 패키지 이름 dto -> gameresult로 변경
