+++ src/main/java/blackjack/domain/BlackJack.java
@@ -2,8 +2,8 @@ package blackjack.domain;
 
 import blackjack.domain.card.CardGenerator;
 import blackjack.domain.card.Deck;
-import blackjack.domain.dto.DistributeResult;
-import blackjack.domain.dto.ProfitResult;
+import blackjack.domain.gameresult.DistributeResult;
+import blackjack.domain.gameresult.ProfitResult;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Participants;
