+++ src/main/java/blackjack/domain/gameresult/ProfitResult.java
@@ -1,4 +1,4 @@
-package blackjack.domain.dto;
+package blackjack.domain.gameresult;
 
 public class ProfitResult {
